pyproject.toml
[project]
name = "learning-flow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "BushraHussain", email = "usamausman2@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "crewai[tools]>=0.100.1",
    "litellm>=1.59.8",
    "python-dotenv>=1.0.1",
]

[project.scripts]
learning-flow = "learning_flow.blog_flow:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
